+++ amethyst_gltf/src/lib.rs
@@ -15,11 +15,7 @@
 
 use std::{collections::HashMap, ops::Range};
 
-use amethyst_animation::{AnimatablePrefab, SkinnablePrefab};
-use amethyst_assets::{
-    AssetStorage, Handle, Loader, Prefab, PrefabData, PrefabLoaderSystem, PrefabLoaderSystemDesc,
-    ProgressCounter,
-};
+use amethyst_assets::{AssetStorage, Handle, Loader, ProgressCounter};
 use amethyst_core::{
     ecs::prelude::{Entity, Read, ReadExpect, Write, WriteStorage},
     math::{convert, Point3, Vector3},
@@ -27,10 +23,7 @@ use amethyst_core::{
     Named,
 };
 use amethyst_error::Error;
-use amethyst_rendy::{
-    camera::CameraPrefab, formats::mtl::MaterialPrefab, light::LightPrefab,
-    rendy::mesh::MeshBuilder, types::Mesh, visibility::BoundingSphere,
-};
+use amethyst_rendy::{rendy::mesh::MeshBuilder, types::Mesh, visibility::BoundingSphere};
 use derivative::Derivative;
 use serde::{Deserialize, Serialize};
 
@@ -38,69 +31,6 @@ pub use crate::format::GltfSceneFormat;
 
 mod error;
 mod format;
-
-/// Builds a `GltfSceneLoaderSystem`.
-pub type GltfSceneLoaderSystemDesc = PrefabLoaderSystemDesc<GltfPrefab>;
-
-/// Load `GltfSceneAsset`s
-pub type GltfSceneLoaderSystem = PrefabLoaderSystem<GltfPrefab>;
-
-/// Gltf scene asset as returned by the `GltfSceneFormat`
-pub type GltfSceneAsset = Prefab<GltfPrefab>;
-
-/// `PrefabData` for loading Gltf files.
-#[derive(Debug, Default)]
-pub struct GltfPrefab {
-    /// `Transform` will almost always be placed, the only exception is for the main `Entity` for
-    /// certain scenarios (based on the data in the Gltf file)
-    pub transform: Option<Transform>,
-    /// `Camera` will always be placed
-    pub camera: Option<CameraPrefab>,
-    /// Lights can be added to a prefab with the `KHR_lights_punctual` feature enabled
-    pub light: Option<LightPrefab>,
-    /// `MeshData` is placed on all `Entity`s with graphics primitives
-    pub mesh: Option<MeshBuilder<'static>>,
-    /// Mesh handle after sub asset loading is done
-    pub mesh_handle: Option<Handle<Mesh>>,
-    /// `Material` is placed on all `Entity`s with graphics primitives with material
-    pub material: Option<MaterialPrefab>,
-    /// Loaded animations, if applicable, will always only be placed on the main `Entity`
-    pub animatable: Option<AnimatablePrefab<usize, Transform>>,
-    /// Skin data is placed on `Entity`s involved in the skin, skeleton or graphical primitives
-    /// using the skin
-    pub skinnable: Option<SkinnablePrefab>,
-    /// Node extent
-    pub extent: Option<GltfNodeExtent>,
-    /// Node name
-    pub name: Option<Named>,
-    pub(crate) materials: Option<GltfMaterialSet>,
-    pub(crate) material_id: Option<usize>,
-}
-
-impl GltfPrefab {
-    /// Move the scene so the center of the bounding box is at the given `target` location.
-    pub fn move_to(&mut self, target: Point3<f32>) {
-        if let Some(ref extent) = self.extent {
-            let diff = convert::<_, Vector3<f32>>(target - extent.centroid());
-            *self
-                .transform
-                .get_or_insert_with(Transform::default)
-                .translation_mut() += diff;
-        }
-    }
-
-    /// Scale the scene to a specific max size
-    pub fn scale_to(&mut self, max_distance: f32) {
-        if let Some(ref extent) = self.extent {
-            let distance = extent.distance();
-            let max = distance.x.max(distance.y).max(distance.z);
-            let scale: f32 = max_distance / max;
-            self.transform
-                .get_or_insert_with(Transform::default)
-                .set_scale(Vector3::new(scale, scale, scale));
-        }
-    }
-}
 
 /// A GLTF node extent
 #[derive(Clone, Debug)]
@@ -218,109 +148,4 @@ pub struct GltfSceneOptions {
     /// Load the given scene index, if not supplied will either load the default scene (if set),
     /// or the first scene (only if there is only one scene, otherwise an `Error` will be returned).
     pub scene_index: Option<usize>,
-}
-
-type SysDataOf<'a, T> = <T as PrefabData<'a>>::SystemData;
-
-impl<'a> PrefabData<'a> for GltfPrefab {
-    // Looks better when defined all in one place
-    #[allow(clippy::type_complexity)]
-    type SystemData = (
-        SysDataOf<'a, Transform>,
-        SysDataOf<'a, Named>,
-        SysDataOf<'a, CameraPrefab>,
-        SysDataOf<'a, LightPrefab>,
-        SysDataOf<'a, MaterialPrefab>,
-        SysDataOf<'a, AnimatablePrefab<usize, Transform>>,
-        SysDataOf<'a, SkinnablePrefab>,
-        WriteStorage<'a, BoundingSphere>,
-        WriteStorage<'a, Handle<Mesh>>,
-        Read<'a, AssetStorage<Mesh>>,
-        ReadExpect<'a, Loader>,
-        Write<'a, GltfMaterialSet>,
-    );
-    type Result = ();
-
-    fn add_to_entity(
-        &self,
-        entity: Entity,
-        system_data: &mut Self::SystemData,
-        entities: &[Entity],
-        children: &[Entity],
-    ) -> Result<(), Error> {
-        let (
-            transforms,
-            names,
-            cameras,
-            lights,
-            materials,
-            animatables,
-            skinnables,
-            bound,
-            meshes,
-            _,
-            _,
-            _,
-        ) = system_data;
-        if let Some(transform) = &self.transform {
-            transform.add_to_entity(entity, transforms, entities, children)?;
-        }
-        if let Some(mesh) = &self.mesh_handle {
-            meshes.insert(entity, mesh.clone())?;
-        }
-        if let Some(camera) = &self.camera {
-            camera.add_to_entity(entity, cameras, entities, children)?;
-        }
-        if let Some(light) = &self.light {
-            light.add_to_entity(entity, lights, entities, children)?;
-        }
-        if let Some(name) = &self.name {
-            name.add_to_entity(entity, names, entities, children)?;
-        }
-        if let Some(material) = &self.material {
-            material.add_to_entity(entity, materials, entities, children)?;
-        }
-        if let Some(animatable) = &self.animatable {
-            animatable.add_to_entity(entity, animatables, entities, children)?;
-        }
-        if let Some(skinnable) = &self.skinnable {
-            skinnable.add_to_entity(entity, skinnables, entities, children)?;
-        }
-        if let Some(extent) = &self.extent {
-            bound.insert(entity, extent.clone().into())?;
-        }
-        Ok(())
-    }
-
-    fn load_sub_assets(
-        &mut self,
-        progress: &mut ProgressCounter,
-        system_data: &mut Self::SystemData,
-    ) -> Result<bool, Error> {
-        let (_, _, _, _, materials, animatables, _, _, _, meshes_storage, loader, mat_set) =
-            system_data;
-
-        let mut ret = false;
-        if let Some(mut mats) = self.materials.take() {
-            mat_set.materials.clear();
-            for (id, mut material) in mats.materials.drain() {
-                ret |= material.load_sub_assets(progress, materials)?;
-                mat_set.materials.insert(id, material);
-            }
-        }
-        if let Some(material_id) = self.material_id {
-            if let Some(mat) = mat_set.materials.get(&material_id) {
-                self.material.replace(mat.clone_loaded());
-            }
-        }
-        if let Some(mesh) = self.mesh.take() {
-            self.mesh_handle =
-                Some(loader.load_from_data(mesh.clone().into(), &mut *progress, meshes_storage));
-            ret = true;
-        }
-        if let Some(animatable) = &mut self.animatable {
-            ret |= animatable.load_sub_assets(progress, animatables)?;
-        }
-        Ok(ret)
-    }
 }